.archieve/xwformats-main/pyproject.toml
# project-src-size=207015 ; 0.20 MB

# Source size (src/*): ~201.0 KB
[build-system]
  requires = ["hatchling"]
  build-backend = "hatchling.build"

[project]
  name = "exonware-xwformats"
  dynamic = ["version"]
  description = "Enterprise serialization formats for xwsystem - Parquet, Avro, HDF5, NetCDF, and 14 more heavyweight formats"
  readme = "README.md"
  requires-python = ">=3.12"
  keywords = []
  classifiers = []
  dependencies = [
    "exonware-xwsystem",  # file-size=2758230 ; 2.63 MB
    # total = src + deps ; 2.83 MB
  ]
  authors = [
    { name = "eXonware Backend Team", email = "connect@exonware.com" },
  ]

  [project.license]
    text = "MIT"

  [project.optional-dependencies]
    lazy = [
      "exonware-xwlazy",  # file-size=349706 ; 0.33 MB
      "exonware-xwsystem[lazy]",  # file-size=3103784 ; 2.96 MB
      # total = src + deps ; 3.49 MB
    ]
    dev = [
      "black>=23.0.0",  # file-size=3914855 ; 3.73 MB
      "exonware-xwlazy",  # file-size=349706 ; 0.33 MB
      "exonware-xwsystem[dev]",  # file-size=40632320 ; 38.75 MB
      "isort>=5.12.0",  # file-size=568232 ; 0.54 MB
      "mypy>=1.0.0",  # file-size=29775254 ; 28.40 MB
      "pytest>=7.0.0",  # file-size=3152415 ; 3.01 MB
      "pytest-asyncio>=0.21.0",  # file-size=46752 ; 0.04 MB
      "pytest-cov>=4.0.0",  # file-size=71639 ; 0.07 MB
      # total = src + deps ; 75.07 MB
    ]
    full = [
      "aiofiles",  # file-size=81991 ; 0.08 MB
      "attrs",  # file-size=440438 ; 0.42 MB
      "attrs2bin>=0.0.1",  # file-size=32002 ; 0.03 MB
      "bcrypt",  # file-size=329770 ; 0.31 MB
      "bottleneck",  # file-size=460187 ; 0.44 MB
      "brotli",  # file-size=868889 ; 0.83 MB
      "cachebox",  # file-size=971831 ; 0.93 MB
      "cachetools",  # file-size=105183 ; 0.10 MB
      "cbor2",  # file-size=406858 ; 0.39 MB
      "charset-normalizer",  # file-size=580726 ; 0.55 MB
      "cloudpickle",  # file-size=125961 ; 0.12 MB
      "cramjam",  # file-size=4216190 ; 4.02 MB
      "cryptography",  # file-size=10265860 ; 9.79 MB
      "cython",  # file-size=13606517 ; 12.98 MB
      "dicttoxml",  # file-size=96721 ; 0.09 MB
      "dill>=0.3.0",  # file-size=822194 ; 0.78 MB
      "exonware-xwsystem[full]",  # file-size=83665879 ; 79.79 MB
      "fastavro>=1.4.0",  # file-size=1404618 ; 1.34 MB
      "flatbuffers>=2.0.0",  # file-size=156467 ; 0.15 MB
      "fsspec",  # file-size=1655173 ; 1.58 MB
      "google",  # file-size=66934 ; 0.06 MB
      "h5py>=3.8.0",  # file-size=8283484 ; 7.90 MB
      "httpx",  # file-size=765633 ; 0.73 MB
      "hyperlight_hyperjson",  # file-size=261754 ; 0.25 MB
      "isal",  # file-size=569869 ; 0.54 MB
      "json5",  # file-size=377396 ; 0.36 MB
      "lmdb>=1.4.0",  # file-size=399680 ; 0.38 MB
      "lz4",  # file-size=277804 ; 0.26 MB
      "msgpack",  # file-size=238387 ; 0.23 MB
      "msgspec",  # file-size=525449 ; 0.50 MB
      "netCDF4>=1.6.0",  # file-size=55631969 ; 53.05 MB
      "numexpr",  # file-size=742653 ; 0.71 MB
      "numpy>=1.20.0",  # file-size=52803492 ; 50.36 MB
      "opentelemetry-api",  # file-size=423339 ; 0.40 MB
      "orjson",  # file-size=289443 ; 0.28 MB
      "pandas>=1.3.0",  # file-size=67223490 ; 64.11 MB
      "pcodec",  # file-size=1021146 ; 0.97 MB
      "psutil",  # file-size=2104166 ; 2.01 MB
      "py-ubjson>=0.16.0",  # file-size=147808 ; 0.14 MB
      "pyarrow>=8.0.0",  # file-size=90047710 ; 85.88 MB
      "pycapnp>=1.1.0",  # file-size=3450539 ; 3.29 MB
      "pylru",  # file-size=74244 ; 0.07 MB
      "pymongo",  # file-size=6208897 ; 5.92 MB
      "pyorc>=0.6.0",  # file-size=2468634 ; 2.35 MB
      "pypostcard>=0.1.0",  # file-size=33211 ; 0.03 MB
      "python-snappy",  # file-size=43189 ; 0.04 MB
      "pytz",  # file-size=1081137 ; 1.03 MB
      "PyYAML",  # file-size=730195 ; 0.70 MB
      "rich",  # file-size=2130726 ; 2.03 MB
      "ron",
      "scipy>=1.7.0",  # file-size=140860699 ; 134.34 MB
      "substrait",  # file-size=1508525 ; 1.44 MB
      "threadpoolctl",  # file-size=125347 ; 0.12 MB
      "thrift>=0.15.0",  # file-size=861988 ; 0.82 MB
      "tomli",  # file-size=288072 ; 0.27 MB
      "tomli-w",  # file-size=25350 ; 0.02 MB
      "xmltodict",  # file-size=65213 ; 0.06 MB
      "zarr>=2.15.0",  # file-size=2847367 ; 2.72 MB
      "zstandard",  # file-size=1384115 ; 1.32 MB
      # total = src + deps ; 540.63 MB
    ]

[tool]
  # Dependency sizes: run 'commands measure_dependency_sizes xwformats' to refresh # project-src-size and # file-size in this file.

  [tool.hatch.version]
    path = "src/exonware/xwformats/version.py"

  [tool.hatch.build]

    [tool.hatch.build.targets]

      [tool.hatch.build.targets.wheel]
        packages = ["src/exonware"]
